1.1.0

+++ Plugin.php
@@ -5,7 +5,7 @@
  * @package UploadGithubForTypecho
  * @author AyagawaSeirin
  * @link https://qwq.best/
- * @version 1.0.3
+ * @version 1.1.0
  * @dependence 1.0-*
  *
  */
@@ -71,13 +71,14 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
                     {
                         document.getElementsByName("desc1")[0].type = "hidden";
                         document.getElementsByName("desc2")[0].type = "hidden";
+                        document.getElementsByName("desc3")[0].type = "hidden";
                         var notice = "正在检查更新...";
                         $.ajax({
                             url: "https://api.github.com/repos/AyagawaSeirin/UploadGithubForTypecho/releases",
                             async: true,
                             type: "GET",
                             success: function (data) {
-                                var now = "1.0.3";
+                                var now = "1.1.0";
                                 var newest = data[0][\'tag_name\'];
                                 if(newest == null){
                                     notice = "检查更新失败，请手动访问插件项目地址获取更新。";
@@ -105,8 +106,9 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
         <li>项目地址：<a href='https://github.com/AyagawaSeirin/UploadGithubForTypecho' target='_blank'>https://github.com/AyagawaSeirin/UploadGithubForTypecho</a><br></li>
         <li>插件使用说明与教程：<a href='https://qwq.best/dev/152.html' target='_blank'>https://qwq.best/dev/152.html</a><br></li>
         <li>插件不会验证配置的正确性，请自行确认配置信息正确，否则不能正常使用。<br></li>
-        <li>插件会替换所有之前上传的文件的链接，若启用插件前存在已上传的文件，请自行将其上传至仓库相同目录中以保证正常显示；同时，禁用插件也会导致链接恢复。插件在上传文件同时也会在本地相应位置留下文件，所以禁用插件后文件链接恢复不用担心文件不存在问题。</li>
+        <li>插件会替换所有之前上传的文件的链接，若启用插件前存在已上传的文件，请自行将其上传至仓库相同目录中以保证正常显示；同时，禁用插件也会导致链接恢复。上传的文件保存在本地的问题请看下面相关配置项。</li>
         <li>注意：由于CDN缓存问题，修改文件后访问链接可能仍然是旧文件，所以建议删掉旧文件再上传新文件，不建议使用修改文件功能。jsDelivr刷新缓存功能暂未推出，推出后本插件会及时更新。</li>
+        <li>Github API限制每个IP每小时只能请求60次接口，请控制您操作图片(上传修改删除)的频率。</li>
         </ol>
         "));
         $github_user = new Typecho_Widget_Helper_Form_Element_Text('githubUser',
@@ -116,8 +118,9 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
         $github_token = new Typecho_Widget_Helper_Form_Element_Text('githubToken', NULL, '', _t('Github账号token'), _t('不知道如何获取账号token请<a href="https://qwq.best/dev/151.html" target="_blank">点击这里</a>'));
         $github_directory = new Typecho_Widget_Helper_Form_Element_Text('githubDirectory',
             NULL, '/usr/uploads', _t('Github仓库内的上传目录'), _t('比如/usr/uploads，最后一位不需要斜杠'));
-        $url_type = new Typecho_Widget_Helper_Form_Element_Select('urlType',
-            array('latest' => '访问最新版本', 'direct' => '直接访问'), 'latest', _t('文件链接访问方式：'), _t('建议选择"访问最新版本"。若修改图片，直接访问方式不方便更新缓存。'));
+        $url_type = new Typecho_Widget_Helper_Form_Element_Select('urlType', array('latest' => '访问最新版本', 'direct' => '直接访问'), 'latest', _t('文件链接访问方式：'), _t('建议选择"访问最新版本"。若修改图片，直接访问方式不方便更新缓存。'));
+        $desc3 = new Typecho_Widget_Helper_Form_Element_Text('desc3', NULL, '', _t('由于Linux权限问题，可能会由于无法创建目录导致文件保存到本地失败而报错异常，请给予本地上传目录777权限。<br>您也可以选择不保存到本地，但可能导致您的主题或其他插件的某些功能异常。<br>您也可以在每一月手动创建当月的目录，避免出现目录创建失败问题（推荐）。'));
+        $if_save = new Typecho_Widget_Helper_Form_Element_Select('ifSave', array('save' => '保存到本地', 'notsave' => '不保存到本地'), 'save', _t('是否保存在本地：'), _t('是否将上传的文件保存在本地。'));
         $desc2 = new Typecho_Widget_Helper_Form_Element_Text('desc2', NULL, '', _t('以下两个参数为选填，留空则为仓库所有者信息。若填写则必须两个都填写。如果您不知道该如何填写，默认即可，不需要修改。'));
         $commit_name = new Typecho_Widget_Helper_Form_Element_Text('commitName', NULL, 'UploadGithubForTypecho', _t('提交文件者名称'), _t('提交Commit的提交者名称，留空则为仓库所属者。'));
         $commit_email = new Typecho_Widget_Helper_Form_Element_Text('commitEmail', NULL, 'UploadGithubForTypecho@typecho.com', _t('提交文件者邮箱'), _t('提交Commit的提交者邮箱，留空则为仓库所属者。'));
@@ -127,6 +130,8 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
         $form->addInput($github_token->addRule('required', _t('请输入Github账号token')));
         $form->addInput($github_directory->addRule('required', _t('请输入Github上传目录')));
         $form->addInput($url_type);
+        $form->addInput($desc3);
+        $form->addInput($if_save);
         $form->addInput($desc2);
         $form->addInput($commit_name);
         $form->addInput($commit_email);
@@ -162,6 +167,8 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
             return false;
         }
         $fileContent = file_get_contents($uploadfile);
+        //echo "fileDir:$fileDir;fileDir_relatively:$fileDir_relatively;path:$path;path_relatively:$path_relatively;";
+
 
         /* 上传到Github */
         $data = array(
@@ -187,19 +194,27 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
         curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
         curl_setopt($ch, CURLOPT_POSTFIELDS, json_encode($data));
         $output = curl_exec($ch);
+        $http_code = curl_getinfo($ch, CURLINFO_HTTP_CODE);
         curl_close($ch);
 
-        /* 写到本地文件 */
-        if (!is_dir($fileDir)) {
-            if (self::makeUploadDir($fileDir)) {
-                file_put_contents($path, $fileContent);
-            } else {
-                //目录创建失败，不写文件
-            }
-        } else {
-            file_put_contents($path, $fileContent);
+        if ($http_code != 201) {
+            $output = json_decode($output,true);
+            self::writeErrorLog($path_relatively, "[Github][upload][" . $http_code . "]" . $output['message']);
         }
 
+        /* 写到本地文件 */
+        if ($options->ifSave == 'save') {
+            if (!is_dir($fileDir)) {
+                if (self::makeUploadDir($fileDir)) {
+                    file_put_contents($path, $fileContent);
+                } else {
+                    //文件写入失败，写入错误日志
+                    self::writeErrorLog($path_relatively, "[local]Directory creation failed");
+                }
+            } else {
+                file_put_contents($path, $fileContent);
+            }
+        }
 
         //返回相对存储路径
         return array(
@@ -239,7 +254,7 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
         $fileContent = file_get_contents($uploadfile);
 
         //判断仓库内相对路径
-        $filename = __TYPECHO_ROOT_DIR__ . $content['attachment']->path;//本地文件绝对路径
+        $filename = __TYPECHO_ROOT_DIR__ . $path;//本地文件绝对路径
         $github_path = $options->githubDirectory . str_replace(self::getUploadDir(), "", $content['attachment']->path);
 
         //获取文件sha
@@ -259,7 +274,7 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
 
         /* 更新Github仓库内文件 */
         $data = array(
-            "message" => "Update file " . str_replace(self::getUploadDir(), "", $content['attachment']->path),
+            "message" => "Modify file " . str_replace(self::getUploadDir(), "", $content['attachment']->path),
             "content" => base64_encode($fileContent),
             "sha" => $sha,
         );
@@ -282,12 +297,21 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
         curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
         curl_setopt($ch, CURLOPT_POSTFIELDS, json_encode($data));
         $output = curl_exec($ch);
+        $http_code = curl_getinfo($ch, CURLINFO_HTTP_CODE);
         curl_close($ch);
-//        print_r($output);
+
+        if ($http_code != 200) {
+            $output = json_decode($output,true);
+            self::writeErrorLog($github_path, "[Github][modify][" . $http_code . "]" . $output['message']);
+        }
 
         //开始处理本地的文件
-        unlink($filename);
-        file_put_contents($filename, $fileContent);
+        if ($options->ifSave == 'save') {
+            if (file_exists($filename)) {
+                unlink($filename);
+            }
+            file_put_contents($filename, $fileContent);
+        }
 
         if (!isset($file['size'])) {
             $file['size'] = filesize($path);
@@ -354,10 +378,17 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
         curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
         curl_setopt($ch, CURLOPT_POSTFIELDS, json_encode($data));
         $output = curl_exec($ch);
+        $http_code = curl_getinfo($ch, CURLINFO_HTTP_CODE);
         curl_close($ch);
 
+        if ($http_code != 200) {
+            $output = json_decode($output,true);
+            self::writeErrorLog($github_path, "[Github][delete][" . $http_code . "]" . $output['message']);
+        }
         //删除本地文件
-        unlink($filename);
+        if ($options->ifSave == 'save' && file_exists($filename) == true) {
+            unlink($filename);
+        }
 
         return true;
     }
@@ -369,7 +400,7 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
     {
         $options = Typecho_Widget::widget('Widget_Options')->plugin('UploadGithubForTypecho');
         $filePath = "https://cdn.jsdelivr.net/gh/" . $options->githubUser . "/" . $options->githubRepo . "@latest" . $content['attachment']->path;
-        return file_get_contents($filePath);//getUploadDir()
+        return file_get_contents($filePath);
     }
 
 
@@ -387,10 +418,23 @@ class UploadGithubForTypecho_Plugin implements Typecho_Plugin_Interface
         return Typecho_Common::url($content['attachment']->path, "https://cdn.jsdelivr.net/gh/" . $options->githubUser . "/" . $options->githubRepo . $latest);
     }
 
+    private static function writeErrorLog($path, $content)
+    {
+        $date = date('[Y/m/d H:i:s]', time());
+        $text = $date . " " . $path . " " . $content . "\n";
+        $log_file = dirname(__FILE__) . "/log/error.log";
+        if (!file_exists($log_file)) {
+            $file = fopen($log_file, 'w');
+        } else {
+            $file = fopen($log_file, 'ab+');
+        }
+        fwrite($file, $text);
+    }
+
     /**
      * 获取安全的文件名
      */
-    private static function getSafeName(&$name)
+    private static function getSafeName($name)
     {
         $name = str_replace(array('"', '<', '>'), '', $name);
         $name = str_replace('\\', '/', $name);
